requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.3.6
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
python-dotenv==1.0.0
Werkzeug==3.0.1
qrcode[pil]==7.4.2
email-validator==2.1.0

